Import FORCIS data

These functions import one specific csv file of the FORCIS database (see below) stored in the folder path. If the csv file does not exist it will be downloaded from Zenodo (https://zenodo.org/record/7936568).

get_plankton_nets_data(path, version = forcis_db_version())

get_pump_data(path, version = forcis_db_version())

get_cpr_north_data(path, version = forcis_db_version())

get_cpr_south_data(path, version = forcis_db_version())

get_sediment_trap_data(path, version = forcis_db_version())

Arguments

path

a character of length 1. The folder in which the FORCIS database has been (or will be) saved.

version

a character of length 1. The version number of the FORCIS database. Default is the latest version.

Value

A data.frame. See https://zenodo.org/record/7390792 for a preview of the dataset.

Details

get_plankton_nets_data() imports the FORCIS plankton nets data

get_pump_data() imports the FORCIS pump data

get_cpr_north_data() imports the FORCIS CPR North data

get_cpr_south_data() imports the FORCIS CPR South data

get_sediment_trap_data() imports the FORCIS sediment traps data

See also