Download the entire FORCIS database

Downloads the entire FORCIS database as a collection of five csv files from Zenodo (https://zenodo.org/record/7936568). Additional files are also downloaded (see https://zenodo.org/record/7936568).

Usage

download_forcis_db(
  path = ".",
  version = options()$forcis_version,
  check_for_update = options()$check_for_update,
  overwrite = FALSE,
  timeout = 60
)

Arguments

path

a character of length 1. The folder in which the FORCIS database will be saved. Note that a subdirectory will be created, e.g. forcis-db/version-99/ (with 99 the version number).

version

a character of length 1. The version number (with two numbers, e.g. 08 instead of 8) of the FORCIS database to download. Default is the latest version. Note that this argument can be handle with the global option forcis_version. For example, if user calls options(forcis_version = "07"), the version 07 will be used by default for the current R session. It is recommended to use the latest version of the database.

check_for_update

a logical. If TRUE (default) the function will check if a newer version of the FORCIS database is available on Zenodo and ask user to download it. Note that this argument can be handle with the global option check_for_update. For example, if user calls options(check_for_update = FALSE), invitation to download the latest version will be disable for the current R session.

overwrite

a logical. If TRUE it will override the downloaded files of the FORCIS database. Default is FALSE.

timeout

a integer. The timeout for downloading files from the FORCIS database. Default is 60. This number can be increased for low Internet connection.

Value

No return value. The FORCIS files will be saved in the path folder.

Examples

if (FALSE) {
# Folder in which the database will be saved ----
path_to_save_db <- "data"

# Download the database ----
download_forcis_db(path = path_to_save_db)

# Check the content of the folder ----
list.files(path_to_save_db, recursive = TRUE)
}